집에서 삭제 기능까지 완료

+++ src/main/java/com/mysite/expensemanager/dto/ExpenseDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.Date;
 
 @Data
@@ -20,7 +22,7 @@ public class ExpenseDTO {
 
     private String description;
 
-    private long amount;
+    private BigDecimal amount;
 
     private Date date;
